+++ src/ui-store.c
@@ -304,14 +304,14 @@ static void store_display_entry(struct menu *menu, int oid, bool cursor, int row
 		x = price_item(store, obj, false, 1);
 
 		/* Make sure the player can afford it */
-		if ((int) player->au < (int) x)
+		if (player->au < x)
 			colour = curs_attrs[CURS_UNKNOWN][(int)cursor];
 
 		/* Actually draw the price */
 		if (tval_can_have_charges(obj) && (obj->number > 1))
-			strnfmt(out_val, sizeof out_val, "%9d avg", x);
+			strnfmt(out_val, sizeof out_val, "%9ld avg", (long)x);
 		else
-			strnfmt(out_val, sizeof out_val, "%9d    ", x);
+			strnfmt(out_val, sizeof out_val, "%9ld    ", (long)x);
 
 		c_put_str(colour, out_val, row, ctx->scr_places_x[LOC_PRICE]);
 	}
@@ -350,8 +350,8 @@ static void store_display_frame(struct store_context *ctx)
 		put_str(owner_name, ctx->scr_places_y[LOC_OWNER], 1);
 
 		/* Show the max price in the store (above prices) */
-		strnfmt(buf, sizeof(buf), "%s (%d)", store_name,
-				proprietor->max_cost);
+		strnfmt(buf, sizeof(buf), "%s (%ld)", store_name,
+				(long)proprietor->max_cost);
 		prt(buf, ctx->scr_places_y[LOC_OWNER],
 			ctx->scr_places_x[LOC_OWNER] - strlen(buf));
 
@@ -450,7 +450,7 @@ static void store_redraw(struct store_context *ctx)
 	}
 
 	if (ctx->flags & (STORE_GOLD_CHANGE)) {
-		prt(format("Gold Remaining: %9d", player->au),
+		prt(format("Gold Remaining: %9ld", (long)player->au),
 				ctx->scr_places_y[LOC_AU], ctx->scr_places_x[LOC_AU]);
 		ctx->flags &= ~(STORE_GOLD_CHANGE);
 	}
@@ -551,14 +551,14 @@ static bool store_sell(struct store_context *ctx)
 	/* Real store */
 	if (store->feat != FEAT_HOME) {
 		/* Extract the value of the items */
-		uint32_t price = price_item(store, temp_obj, true, amt);
+		int32_t price = price_item(store, temp_obj, true, amt);
 
 		object_wipe(temp_obj);
 		screen_save();
 
 		/* Show price */
 		if (!OPT(player, birth_no_selling))
-			prt(format("Price: %d", price), 1, 0);
+			prt(format("Price: %ld", (long)price), 1, 0);
 
 		/* Confirm sale */
 		if (!store_get_check(format("%s %s? [ESC, any other key to accept]",
@@ -616,7 +616,7 @@ static bool store_purchase(struct store_context *ctx, int item, bool single)
 
 		/* Check if the player can afford any at all */
 		if (store->feat != FEAT_HOME &&
-				(int)player->au < (int)price_item(store, obj, false, 1)) {
+				player->au < price_item(store, obj, false, 1)) {
 			msg("You do not have enough gold for this item.");
 			return false;
 		}
@@ -708,7 +708,7 @@ static bool store_purchase(struct store_context *ctx, int item, bool single)
 		screen_save();
 
 		/* Show price */
-		prt(format("Price: %d", price), 1, 0);
+		prt(format("Price: %ld", (long)price), 1, 0);
 
 		/* Confirm purchase */
 		response = store_get_check(format("Buy %s?%s %s",
